ref: move config package to configuration
