+++ src/main/java/fr/insee/pogues/metadata/client/MetadataClientImpl.java
@@ -1,24 +1,21 @@
 package fr.insee.pogues.metadata.client;
 
-import java.util.Arrays;
-import java.util.List;
-
-import org.apache.http.entity.ContentType;
+import fr.insee.pogues.configuration.RemoteMetadata;
+import fr.insee.pogues.metadata.model.ColecticaItem;
+import fr.insee.pogues.metadata.model.ColecticaItemRefList;
+import fr.insee.pogues.metadata.model.Unit;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.web.client.RestTemplateBuilder;
-import org.springframework.http.HttpEntity;
-import org.springframework.http.HttpMethod;
-import org.springframework.http.ResponseEntity;
+import org.springframework.http.*;
 import org.springframework.stereotype.Service;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
 
-import fr.insee.pogues.config.RemoteMetadata;
-import fr.insee.pogues.metadata.model.ColecticaItem;
-import fr.insee.pogues.metadata.model.ColecticaItemRefList;
-import fr.insee.pogues.metadata.model.Unit;
+import java.util.Arrays;
+import java.util.List;
 
+// TODO: change for webclient
 @Service
 public class MetadataClientImpl implements MetadataClient {
 
@@ -42,7 +39,7 @@ public class MetadataClientImpl implements MetadataClient {
     public List<ColecticaItem> getItems(ColecticaItemRefList query) throws Exception {
         String url = String.format("%s/meta-data/items", remoteMetadata.getUrl());
         MultiValueMap<String, String> headers = new LinkedMultiValueMap<>();
-        headers.add("Content-type", ContentType.APPLICATION_JSON.getMimeType());
+        headers.add(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE);
         HttpEntity<ColecticaItemRefList> request = new HttpEntity<>(query, headers);
         ResponseEntity<ColecticaItem[]> response = restTemplate
                 .exchange(url, HttpMethod.POST, request, ColecticaItem[].class);
